# apps/desktop/src/app/core/api/model/album.ts
/**
 * Lifeframe Sync & Media API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Album { 
    id: string;
    name: string;
    relative_path: string;
    photo_count: number;
}