# apps/desktop/src/app/core/api/model/device-status-response.ts
/**
 * Lifeframe Sync & Media API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DeviceStatusResponse { 
    success: boolean;
    message: string;
    server_time: string;
}